engine: integer floor-division for div per spec §07
Per §07, div(a, b) = floor(a/b) over integers; real/complex division is
the separate divide op. div was typed (real, real) -> real and evaluated
as true division. Type it (int, int) -> int, floor it across every path
(value-ops floorDiv/floorDivElem, scalar sampler, materialiser fold,
variant/broadcast dispatch, and the ir-shared constant fold), and make a
real or complex operand a typecheck error. divide is unchanged.

Breaking: div on real/complex operands is now a typecheck error, and
integer div now floors instead of truncating toward zero. Use divide for
real division. No in-tree callers affected.

+++ packages/engine/types.ts
@@ -852,7 +852,11 @@ const SIGNATURE_FACTORIES = {
   add:    () => arith2(),
   sub:    () => arith2(),
   mul:    () => arith2(),
-  div:    () => ({ args: [REAL, REAL], kwargs: {}, result: REAL }),  // div always real
+  // div(a, b) = floor(a/b), spec §07 (line 419): domain `integers`,
+  // `b ≠ 0`, result integer (integer floor-division, NOT real division).
+  // Real/complex division is the separate `divide` op (line 449), still
+  // typed (real, real) → real below — div ≠ divide.
+  div:    () => ({ args: [INTEGER, INTEGER], kwargs: {}, result: INTEGER }),
   divide: () => ({ args: [REAL, REAL], kwargs: {}, result: REAL }),
   // mod(a, b) = a − b·floor(a/b), spec §07: domain `integers`, `b ≠ 0`,
   // result integer (floor-modulo, NOT JS `%` truncated remainder). The
